# contracts/openapi/platform.yaml
openapi: 3.0.3
info:
  title: BuildingOS platform foundation API
  version: "0.1.0"
  description: >
    Exact operational routes for the BOS-001 platform foundation milestone.
    Service metadata plus BOS-010 F5a revenue APIs (subscription-service) and F2 building
    application & lifecycle APIs (building-service).
servers:
  - url: http://localhost:8080
    description: Local gateway
paths:
...
  /api/v1/platform/buildings/{buildingId}:
    get:
      summary: "Building with memberships"
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "BUILDING_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/buildings/{buildingId}/activate:
    post:
      summary: "ONBOARDING to ACTIVE (needs an active BUILDING_ADMIN, interim D-30)"
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (reason required)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "BUILDING_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION or NO_BUILDING_ADMIN"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/buildings/{buildingId}/suspend:
    post:
      summary: "ACTIVE to SUSPENDED (deletes nothing)"
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (reason required)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "BUILDING_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /api/v1/platform/buildings/{buildingId}/reactivate:
    post:
      summary: "SUSPENDED to ACTIVE"
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReasonRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (reason required)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (requires SUPER_ADMIN or PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: "BUILDING_NOT_FOUND (also for applications the caller does not own)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: "INVALID_TRANSITION"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /internal/users/provision:
    post:
      summary: "auth-service: find or create a user by phone (service-to-service, D-29)"
      description: >
        Called by building-service with the approver's relayed token. Never routed by the gateway. Grants no roles.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProvisionUserRequest" }
      responses:
        "200":
          description: "Success envelope"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiEnvelope" }
        "400":
          description: "INVALID_REQUEST (not a Bangladesh mobile number)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401":
          description: "Missing or invalid token"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: "ACCESS_DENIED (relayed token lacks SUPER_ADMIN/PLATFORM_ADMIN)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, traceId]
      properties:
        success: { type: boolean }
        data: {}
        meta: { type: object }
        traceId: { type: string }
...
    ReasonRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, maxLength: 1000 }
...
    ProvisionUserRequest:
      type: object
      required: [phone]
      properties:
        phone: { type: string }
    ApiError:
      type: object
      required: [success, code, message, traceId]
      properties:
        success: { type: boolean, enum: [false] }
        code:
          type: string
          description: Stable error code; each operation lists the codes it may return.
          example: ACCESS_DENIED
        message: { type: string }
        traceId: { type: string, format: uuid }